Resolve Resonator amount not being reduced when broken.
Usage cases now covered:

* TNT explosions in both survival & creative
* Breaking by hand in creative
* Breaking with an axe in survival (both specifically drop crystals)
* Destruction by another entity (Creeper Explosion, Wither Explosion, Dragon)

I unified the code for all of these into onBlockDestroyed, which is
called by the various events that can destroy blocks. While explosions
and by the player is covered by the default functions of Block, by other
entities is not.

Additionally, added GetNetworkID to the to instances of
ArcaneArchivesNetwork to prevent directly accessing the NetworkID.

It would be worth considering extending this to cover any instance of
BlockTemplate (excluding radiant chests), so that future tile entities
which have their blocks destroyed are also cleared from mNetworkBlocks.