Move event handlers into proxy for future Client/Server separation.
Also delineate the variance of Instances of the mod. Now there is only
the one, public instance available at ArcaneArchives.instance. The
instance() function servers no purpose that I can see, and it referred
to INSTANCE, which was instantiated separately.

instance is automatically instantiated by Forge when appropriate and is
the correct reference to use.