Turned on Lint warnings. Fixed a number of unchecked assignments.
ArrayList() does actually need to be ArrayList<>() for some compilers
and platforms and it's just good practice. Also converted a few instaces
of Arrays.asList to Lists.newList as the former doesn't /actually/
return an ArrayList or List instance (which the function was expecting).

Removed a number of new assignations of member Lists, etc, from being
myMember = new ArrayList<>() and simply replaced them with .clear()
which is functionally equivalent without having the overhead of creating
a new object.

There's at least one instance (Gem Cutter's Table recipe handler) where
there's a presume order to the keySet returned by a Map. Sets are
generally orderless and to get an index based on the slot number could
potentially return wildly different results, as the order in which items
come out of the Set and are added to the ArrayList could change
depending on platform, implementation, etc.