Final block of changes: calculateAccessors, onBlockPlaceBy.
The majority of the TileEntity modifying code should actually be in
onBlockPlaceBy, however the ItemBlockTemplate still retains the code for
actually failing to place a block when the network limit is reached.

Speaking of network limits, for some reason the first placement of a
Matrix core in a world results in it being placed properly. But
subsequent placements, tested with breakpoints on the actual block code
that creates the new tile entity, prove that a new tile entity can be
placed before the onItemUse code is ever consulted.

This resulted in a situation where the network *did* find a relevant
tile entity and thus decided it was over the place limit, but that the
tile entity did not exist.

If this is a bug, hopefully it can be resolved in the future. For the
moment, wrapped with a snuggly TODO, I've split it out into fetching all
relevant tile entities, and, for placeLimits of 1, testing the last
entry in the list to make sure that it isn't directly above where the
player was just targetting -- if it is, then it's the new tile entity
and it doesn't count.

Annoying!

Finally, code in the onBlockPlacedBy is currently firing to place stone
blocks as pseudo-accessor blocks. Suitable accessor blocks and their
relevant tile entities will be added at a later date, but for now it's
enough to visualise the actual shape.

Keeping in mind that, while the accessor blocks should adjust properly
to facing, if the model itself doesn't have a facing entry, it will
render strangely offset (Matrix Reservoir primarily).