Refactor BlockLibrary slightly, handle IDs for BlockTemplates.
An integer property is the only sane way for Accessors to know what
block they're actually acting as an accessor of. Thus, I've arbitrarily
chosen the order that they were listed in the BlockLibrary as the order
increasing from 0.

I've moved the tile entity list creation code into the BlockLibrary as
it makes more sense for it to be here than in the RegistryHandler. I've
simplified the code slightly and refactored out a foreach.

It didn't make sense to have indexes set by the order upon which blocks
were added (by their instantiator) to be the integer type of the block
as there's a (slight) possibility that this order might change.

In order to make things simpler, I used a HashBiMap to store the
Integer<->Block comparisons. This makes it easier for the accessor
placement code to specify which block it is being placed for (by calling
getAccessorId()), without having to maintain two separate maps.