Implement Accessor Blocks.
Basically, these pass through pretty much everything that's handled in
the State to the base block. This presumes that they are correctly given
a TYPE value when created.

Currently unhandled: breaking one of the accessor blocks should have the
effect of breaking the entire structure, but as I haven't actually set
the code up to use accessor yet, that cane come later.