Replicate facing from BlockDirectionalTemplate::getStateForPlacement.
Seeing as this facing value (well, the opposite thereof) is what the
accessor blocks are using when they're placed, it makes sense to use
that mechanism instead of the get-entity-facing-pos etc, which most of
the time points down.

This brings both instances of calculateAccessors into line with each
other, and prevents you from placing, say, Gem Cutter's Tables into pure
stone.