So... it's like when you start organising one drawer, and you...
... take it out, and then you take out another. Then you start looking
before another drawer. Then you decide that you want to organise things
in a slightly different way, and then you realise that all of what
you've done has had absolutely no bearing on the issue.

The issue at hand is the following: ItemStack may be equal but do not
have the same .hashCode(), meaning they cannot be used as indexes for
HashMaps.

The solution to all our problems was refactoring the GemCutter's Recipe
and Recipe List, which are now separate classes in order to reduce
confusion.

In addition, I removed the changed recipe packet because that's one that
can come from the server-side of the container object. However, the
change page packet still seems to be required.

I also rewrote the Item Consolidation, removed the GCT Item Handler in
favor of a straight ItemStackHandler for the inputs/output slot
(although the latter needs to be "locked"), and then added
SlotRecipeHandler to handle the actual recipes instead -- that way if we
ever end up with more recipes in the table.

I should've gone to bed ages ago.