Let's do it all in one big lump.
ArcaneArchivesNetwork is now server-only. The client-side copy is passed
via packet and accesible via the NetworkHelper but via a different
method.

Currently, tiles are not passed to the client network, simply the item
handler. The item handler should be modified as it doesn't deal with
chests properly.

I've marked (for the time-being, I know this isn't how it's supposed to
be used) the various functions with @SideOnly so we can get crashes
whenever one of them gets called in the wrong thread.

This is for testing purposes only.