Generify ReferenceList, make ManifestList. Complete ManifestItemHandler.
It now directly referencess a filtered/unfiltered view of a list instead
of creating more lists. The container itself will need to be updated,
but as far as I'm aware, it can be set up to do nothing on the server
side.

Once the container is functional, it's a matter of linking the search
term back into the ManifestList from the GUI, making sure the separate
stacks are being displayed per dimension, and then, if these aren't the
player's current dimension, setting the background of the slot to a
relevant colour.

Grey for non-standard Minecraft dimensions, deep red for Nether, yellow
for the End, perhaps?