Temporarily convert radiant chest into a 54-slot chest.
The size may increase but packet issues and the lack of time to
implement our own GUI make it basically impossible to implement this
properly.

While it functions perfectly in single-player, and while enough packets
could be overriden to make its functionality work, the moment another
mod introduces a container listener (for capabilities or whatsoever),
there's the chance that the same issues could start up again.

Changes to the radiant chest may come in a future version!