Allow invalid ITEs (placed by Rannucarpus for example) to tick... (Brian)
Once they've ticked for long enough (currently 40 ticks) to compensate
for the fact that it might just be a bit of lag with the
onBlockPlacedBy code (which theoretically should be within the next tick
if not the same tick), they'll simply be destroyed if they aren't part
of a valid network.

This resolves the issue that Brian reported on the Discord using the
Rannucarpus to theoretically bypass the limits of the radiant
resonators.

(In fact, they were simply being placed for nobody and were also
bypassing all of the item-use code in the ItemBlockTemplate).