Considerable amount of work on the Gem Cutter's Table.
I can't even begin to summarise it. Recipes are now derived from
shapeless recipes, although there is actually issues with crafting at
the minute due to the ... I don't know what, but there's item
duplication GALORE!

Also, things are out of order with regards to the GCT recipe list. This
will be fixed.