HWYLA/WAILA, TOP, Quark & InventoryTweaks support.
Unfortunately to actually get it to compile in a dev environment without
triggering access violations (even with a deobfuscated jar), I had to
merge their access transformers into our own.

I'm going to leave them as-is for now but presumably they can be removed
at some point in the future when we don't need to keep testing it.