Fix issues with shift-clicking into the GCT.
Output is now, by default, only sent into the player inventory. If the
player inventory is full and there are slots available in the tile
inventory, *it will fail*.

This hopefully resolves issues where items "disappear" when trying to
shift-click them into the tile. (Simply because it was attempting to
shift click them into the output slot, where they just got eaten.)