The Saga Of The Mixin draws to a close.
Basically, after 8 or so hours we now get pretty backgrounds in non-mod
containers. Hooray!

I also moved the invtweaks API code into the main source module as
IntelliJ had issues with it and kept insisting that it no longer
existed. This was not the only issue I had with IntelliJ, and it appears
the Java Curse is now spreading.