A more agnostic and hands-off approach to duplicated monitoring crystals.
Instead of the additional O(n)*<number of monitoring crystals> (I
think), this implementation should be the O of ArrayList::contains. Sets
might be faster but I'm not sure it matters at this point.

Either way, this is agnostic in that it merely checks to see if the
block position has already been indexed, then indexes itself, rather
than consulting every other monitoring crystal.