Potentially resovle IndexError with server->client overwrites.
The server attempts to enforce its own idea of what should be in the
slots of the container which is purely client-side and which the server,
frankly, doesn't need to know about.

Hopefully this fixes the "race condition" of the container being
created server-side and injected with the player as an
IContainerListener and sendAllContents being triggered immediately,
before the container is finished being initialised on the client-side.