Rewrite chest->radiant chest conversion process to be more robust.
It will no longer automatially assume that any adjacent instance of a
BlockChest nearby is actually part of it as a double chest (removing
issues where birch/etc chests from Quark were completely broken).

In fact, it now just fetches the "double chest handler" and consumes
from that: there is no longer any code to check for adjacency -- all of
that is left up to the vanilla/etc implementation of the double chest.