Allow for chests to have items displayed on any horizontal face.
This also includes blocks. Items can be displayed by holding a scepter
of Manipulation in one hand and the item in the other, then right
clicking on the other face.

Items can be cleared by sneak-right-clicking with a scepter of
manipulation and an empty offhand (or main hand/vice versa).