Resolve #100: properly remove broken monitoring tiles.
It appears that the code in the block-breaking system for these was
either never implemented properly or was faulty; now you don't need to
log out and back in again to prevent spam when using a manifest.

However, as it currently stands, it only looks at block positions as
there's no real way to determine a unique identity for an inventory.
Thus, it's possible to put two monitoring crystals on a double chest.

I'll be writing a work-around for this.