Finish cleaning up ManifestItemHandler.
Filtering is now an intrinsic part of sorting. Sorting always creates a
new copy of the list (it is not possible to create a sorting iterator).
Likewise, as the ManifestItemHandler requires per-slot access, a
filtered iterator is insufficient.

As the manifest list is never consulted without being sorted first,
combining filtering and sorting into one method makes logical sense.

Thus, if filters are evident, apply them before returning the sorted
copy.