Genericise processors a little more.
Now a processor accepts the result item (which can be modified in the
return value), a list of ingredients, a list of a list of itemstacks
that make up the result (in case processors give multiple results,
perhaps), as well as access to the ICrafter.

Moved away from locking it to WorkbenchCrafting because of the potential
for other types of crafting, including the in-world keystone.